예제3 - useInput 커스텀 훅을 생성해서 컴포넌트 별 중복 코드 제거
